+++ back_end/bug_app/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
-from .models import Project, Sprint, Ticket, Company
+from .models import Project, Sprint, Ticket
+from company_app.models import Company
 from user_app.models import Client
 
 # Register your models here.
