+++ back_end/bug_tracker_project/urls.py
@@ -20,5 +20,5 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/users/', include('user_app.urls')),
-    path('api/companies/', include('bug_app.urls'))
+    path('api/companies/', include('company_app.urls'))
 ]
