+++ back_end/company_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CompanyAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'company_app'
