+++ back_end/company_app/models.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+# Create your models here.
+class Company(models.Model):
+    name = models.CharField(max_length=50, unique=True)
+
+    def __str__(self):
+        return self.name
+
